dcs: Update to use new API
