+++ pkg/qrcodegenerator/qrcode/enddevices/enddevices.go
@@ -1,0 +1,110 @@
+// Copyright © 2022 The Things Network Foundation, The Things Industries B.V.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package enddevices
+
+import (
+	"context"
+	"fmt"
+	"sync"
+
+	"go.thethings.network/lorawan-stack/v3/pkg/errors"
+	"go.thethings.network/lorawan-stack/v3/pkg/qrcodegenerator/qrcode"
+	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+)
+
+var (
+	errCharacter = errors.DefineInvalidArgument("character", "invalid character `{r}`")
+	errNoJoinEUI = errors.DefineFailedPrecondition("no_join_eui", "no JoinEUI")
+	errNoDevEUI  = errors.DefineFailedPrecondition("no_dev_eui", "no DevEUI")
+	errFormat    = errors.DefineInvalidArgument("format", "invalid format")
+)
+
+// Format is a end device QR code format.
+type Format interface {
+	Format() *ttnpb.QRCodeFormat
+	New() Data
+	// Return the ID of this format as a string.
+	ID() string
+}
+
+// Data represents end device QR code data.
+type Data interface {
+	qrcode.Data
+	Encode(*ttnpb.EndDevice) error
+}
+
+// GetEndDeviceFormats returns the registered end device QR code formats.
+func (c *QRCode) GetEndDeviceFormats() map[string]Format {
+	res := make(map[string]Format)
+	c.endDeviceFormatsMu.RLock()
+	for k, v := range c.endDeviceFormats {
+		res[k] = v
+	}
+	c.endDeviceFormatsMu.RUnlock()
+	return res
+}
+
+// GetEndDeviceFormat returns the converter by ID.
+func (c *QRCode) GetEndDeviceFormat(id string) Format {
+	c.endDeviceFormatsMu.RLock()
+	res := c.endDeviceFormats[id]
+	c.endDeviceFormatsMu.RUnlock()
+	return res
+}
+
+// RegisterEndDeviceFormat registers the given end device QR code format.
+// Existing registrations with the same ID will be overwritten.
+func (c *QRCode) RegisterEndDeviceFormat(id string, f Format) {
+	c.endDeviceFormatsMu.Lock()
+	c.endDeviceFormats[id] = f
+	c.endDeviceFormatsMu.Unlock()
+}
+
+// QRCode is a QR Code used for entity onboarding.
+type QRCode struct {
+	endDeviceFormats   map[string]Format
+	endDeviceFormatsMu sync.RWMutex
+}
+
+// New returns a QRCode.
+func New(ctx context.Context) *QRCode {
+	return &QRCode{
+		endDeviceFormats: make(map[string]Format),
+	}
+}
+
+var (
+	errUnknownFormat = errors.DefineInvalidArgument("unknown_format", "format `{format_id}` unknown")
+)
+
+// Parse attempts to parse the given QR code data.
+// It returns the parser and the format ID that successfully parsed the QR code.
+func (c *QRCode) Parse(formatID string, data []byte) (Data, error) {
+	for id, format := range c.endDeviceFormats {
+		// If format ID is provided, use  only that.
+		if formatID != "" && formatID != id {
+			continue
+		}
+		edFormat := format.New()
+		fmt.Println(id, formatID)
+		if err := edFormat.UnmarshalText(data); err == nil {
+			return edFormat, nil
+		} else if formatID == id {
+			// Return the parsing error is format was specified.
+			return nil, err
+		}
+	}
+	return nil, errUnknownFormat.WithAttributes("format_id", formatID)
+}
