+++ pkg/qrcodegenerator/util_test.go
@@ -21,7 +21,7 @@ import (
 
 	pbtypes "github.com/gogo/protobuf/types"
 	"go.thethings.network/lorawan-stack/v3/pkg/component"
-	"go.thethings.network/lorawan-stack/v3/pkg/qrcodegenerator/qrcode"
+	"go.thethings.network/lorawan-stack/v3/pkg/qrcodegenerator/qrcode/enddevices"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/v3/pkg/types"
 )
@@ -51,7 +51,7 @@ func (m *mock) Encode(dev *ttnpb.EndDevice) error {
 	return nil
 }
 
-func (*mock) GetEntityOnboardingData() *ttnpb.EntityOnboardingData { return nil }
+func (*mock) EndDeviceInfo() (string, *ttnpb.EndDeviceTemplate) { return "", nil }
 
 func (m mock) MarshalText() ([]byte, error) {
 	return []byte(fmt.Sprintf("%s:%s", m.ids.JoinEui, m.ids.DevEui)), nil
@@ -71,7 +71,7 @@ func (mockFormat) Format() *ttnpb.QRCodeFormat {
 	}
 }
 
-func (mockFormat) New() qrcode.EndDeviceData {
+func (mockFormat) New() enddevices.Data {
 	return new(mock)
 }
